requirements.txt
alabaster==0.7.12
astroid==2.4.1
Babel==2.8.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
docutils==0.16
file-magic==0.4.0
idna==2.9
imagesize==1.2.0
isort==4.3.21
Jinja2==2.11.2
jsonpointer==2.0
lazy-object-proxy==1.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
mutagen==1.44.0
ntlm-auth==1.4.0
packaging==20.4
promise-keeper==0.4
psycopg2==2.8.5
pyasn1==0.4.8
pycparser==2.20
Pygments==2.6.1
pylint==2.5.2
pyparsing==2.4.7
PyPDF2==1.26.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
six==1.15.0
smbprotocol==1.0.1
snowballstemmer==2.0.0
Sphinx==3.0.4
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.17
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
